Remove predefined margin in Group block
Currently, there is no user setting to set margins for a Group block in WordPress. Having a predefined bottom margin of 20px makes it difficult for users to unset/overide it. If users wanted a space before/after a Group block, they can easily use a Spacer block to define it.