submenu text alignment (see description)
For the customize preview, we use style tag instead of adding wpbf-text-thealignmentvalue because using style tag is simpler. We simply update the style value.

If we use wpbf-text-thealignmentvalue in the customize preview, we will have to select all sub menu then assign the class to those elements.

And in the output, it's also simpler. We simply output the style to css php file.
If we use wpbf-text-thealignmentvalue, then likely we need to hook a function.